Moving the join tables to their own package for better organization
