+++ src/main/java/usi/si/seart/job/CrawlProjectsJob.java
@@ -22,10 +22,10 @@ import usi.si.seart.exception.UnsplittableRangeException;
 import usi.si.seart.github.GitCommit;
 import usi.si.seart.github.GitHubAPIConnector;
 import usi.si.seart.model.GitRepo;
-import usi.si.seart.model.GitRepoLanguage;
 import usi.si.seart.model.Label;
 import usi.si.seart.model.Language;
 import usi.si.seart.model.Topic;
+import usi.si.seart.model.join.GitRepoLanguage;
 import usi.si.seart.service.GitRepoService;
 import usi.si.seart.service.LabelService;
 import usi.si.seart.service.LanguageService;
