+++ src/main/java/usi/si/seart/model/GitRepo.java
@@ -10,6 +10,8 @@ import lombok.experimental.FieldDefaults;
 import org.hibernate.Hibernate;
 import org.hibernate.annotations.Fetch;
 import org.hibernate.annotations.FetchMode;
+import usi.si.seart.model.join.GitRepoLanguage;
+import usi.si.seart.model.join.GitRepoMetric;
 
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
