+++ src/main/java/usi/si/seart/model/join/GitRepoLanguage.java
@@ -1,4 +1,4 @@
-package usi.si.seart.model;
+package usi.si.seart.model.join;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -7,6 +7,9 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.experimental.FieldDefaults;
+import org.hibernate.Hibernate;
+import usi.si.seart.model.GitRepo;
+import usi.si.seart.model.Language;
 
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
@@ -26,9 +29,9 @@ import java.util.Objects;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name = "git_repo_label")
+@Table(name = "git_repo_language")
 @Entity
-public class GitRepoLabel {
+public class GitRepoLanguage {
 
     @EmbeddedId
     Key key;
@@ -39,9 +42,26 @@ public class GitRepoLabel {
     GitRepo repo;
 
     @ManyToOne(optional = false)
-    @MapsId("labelId")
-    @JoinColumn(name = "label_id")
-    Label label;
+    @MapsId("languageId")
+    @JoinColumn(name = "language_id")
+    Language language;
+
+    @NotNull
+    @Column(name = "size_of_code")
+    Long sizeOfCode;
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
+        GitRepoLanguage language = (GitRepoLanguage) o;
+        return getKey() != null && Objects.equals(getKey(), language.getKey());
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(getKey());
+    }
 
     @Getter
     @Setter
@@ -55,20 +75,20 @@ public class GitRepoLabel {
         Long repoId;
 
         @NotNull
-        @Column(name = "label_id")
-        Long labelId;
+        @Column(name = "language_id")
+        Long languageId;
 
         @Override
         public boolean equals(Object obj) {
             if (this == obj) return true;
             if (obj == null || getClass() != obj.getClass()) return false;
             Key other = (Key) obj;
-            return repoId.equals(other.repoId) && labelId.equals(other.labelId);
+            return repoId.equals(other.repoId) && languageId.equals(other.languageId);
         }
 
         @Override
         public int hashCode() {
-            return Objects.hash(repoId, labelId);
+            return Objects.hash(repoId, languageId);
         }
     }
 }
