+++ src/main/java/usi/si/seart/repository/specification/GitRepoSearch.java
@@ -5,11 +5,11 @@ import lombok.Builder;
 import org.apache.commons.lang3.StringUtils;
 import usi.si.seart.model.GitRepo;
 import usi.si.seart.model.GitRepoMetricAggregate_;
-import usi.si.seart.model.GitRepoMetric_;
 import usi.si.seart.model.GitRepo_;
 import usi.si.seart.model.Label_;
 import usi.si.seart.model.Language_;
 import usi.si.seart.model.Topic_;
+import usi.si.seart.model.join.GitRepoMetric_;
 import usi.si.seart.repository.criteria.Criteria;
 import usi.si.seart.repository.criteria.KeyCriteria;
 import usi.si.seart.repository.criteria.KeyValueCriteria;
